Deprecate Operator.validate_subspace() after black
